refactor: enum 정리 #80
