+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/Hi.java
@@ -1,9 +1,7 @@
 package com.gdg.z_meet.domain.meeting.entity;
 
-import com.gdg.z_meet.domain.meeting.entity.status.HiStatus;
+import com.gdg.z_meet.domain.meeting.entity.enums.HiStatus;
 import com.gdg.z_meet.global.common.BaseEntity;
-import com.gdg.z_meet.global.exception.BusinessException;
-import com.gdg.z_meet.global.response.Code;
 import jakarta.persistence.*;
 import lombok.*;
