+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/enums/HiStatus.java
@@ -1,4 +1,4 @@
-package com.gdg.z_meet.domain.meeting.entity.status;
+package com.gdg.z_meet.domain.meeting.entity.enums;
 
 public enum HiStatus {
     NONE, ACCEPT, REFUSE, EXPIRED
