+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/enums/MatchingStatus.java
@@ -1,0 +1,5 @@
+package com.gdg.z_meet.domain.meeting.entity.enums;
+
+public enum MatchingStatus {
+    WAITING, COMPLETE
+}
