+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/enums/TeamType.java
@@ -1,4 +1,4 @@
-package com.gdg.z_meet.domain.meeting.entity;
+package com.gdg.z_meet.domain.meeting.entity.enums;
 
 public enum TeamType {
     TWO_TO_TWO, THREE_TO_THREE
