+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/enums/Verification.java
@@ -1,4 +1,4 @@
-package com.gdg.z_meet.domain.meeting.entity;
+package com.gdg.z_meet.domain.meeting.entity.enums;
 
 public enum Verification {
     NONE, IN_PROGRESS, COMPLETE
