+++ src/main/java/com/gdg/z_meet/domain/meeting/service/HiQueryService.java
@@ -2,7 +2,6 @@ package com.gdg.z_meet.domain.meeting.service;
 
 import com.gdg.z_meet.domain.meeting.dto.MeetingRequestDTO;
 import com.gdg.z_meet.domain.meeting.dto.MeetingResponseDTO;
-import com.gdg.z_meet.domain.meeting.entity.TeamType;
 
 import java.util.List;
