+++ src/main/java/com/gdg/z_meet/domain/meeting/service/MeetingQueryService.java
@@ -1,12 +1,7 @@
 package com.gdg.z_meet.domain.meeting.service;
 
-import com.gdg.z_meet.domain.meeting.dto.MeetingRequestDTO;
 import com.gdg.z_meet.domain.meeting.dto.MeetingResponseDTO;
-import com.gdg.z_meet.domain.meeting.entity.TeamType;
-import com.gdg.z_meet.domain.meeting.entity.status.HiStatus;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import java.util.List;
+import com.gdg.z_meet.domain.meeting.entity.enums.TeamType;
 
 public interface MeetingQueryService {
